feat: add swagger documentation (#94)
### Overview

This PR adds swagger documentation, and make several improvements on the
API.

It uses the ["tsoa" framework](https://github.com/lukeautry/tsoa), which
is a bit like a
[DSL](https://en.wikipedia.org/wiki/Domain-specific_language) (Domain
specific language), and generates both the express routes and the
openapi documentation, based on the typed objects and jsdocs.

### Features 

- adds swagger documentation at the `api/docs` endpoint (using
[`swagger-ui-express`](https://www.npmjs.com/package/swagger-ui-express))
- improved dev environment with unique `npm run dev` that runs all
hot-reloading servers at once (using
[`concurrently`](https://www.npmjs.com/package/concurrently)).
- Better Error handling & errors documented in Swagger

### Refactorings

- Put API logic from "index.ts" into "controller/app.ts", to keep only
server related stuff in "index.ts"
- I was struggling with ES6 module support, so I went back to CommonJS
code generation. As a result, `true-myth` dependency was failing, but I
could make it work thanks to a recent CommonJs compatibility fix (in
v7.0.1). As a result, relative imports do not need ".js".
- "controller/routes.ts" moved to "sireneController.ts" with tsoa
automatic discovery (configured in "tsoa.json")

### Details

- Code generation adds another build step. The generated code goes into
the "generated" folder (gitignored).