fix + upgrade : body correction before sending API request to backend + backend url as env var (#100)
### Problem : 

The frontend was sending a request to the TEE-backend API (the new one
in Express) with body containing more than the only key necessary
(`siret : ...`), resulting in a 422 error from the server.

### Resolution : 

The body sent in the request was built from a common data structure in
`trackSiret.ts`'s callbacks, instead of having a specific body structure
to fill. I added a specific `dataBody` object in this request
configuration file, and added in `request.ts`'s `sendApiRequest`function
an indication to use this dataBody before dataStructure (and added its
corresponding type for TypeScript)

--- 

## Upgrades

### Backend URL

The backend url was hard written in `trackSiret`file. For development
and tests purposes it was decided to migrate this info in the
environment variables (set in Scalingo), so the Scalingo apps :

- `preprod-front` could directly request to `preprod-backend` or to
`prod-backend`
- `prod-front` could directly request to `prod-backend`
- `staging-front` could directly request to any backend (prod | preprod
| staging) we'd like

### Refacto

Some functions were duplicated between `helpers.ts` / `request.ts` and a
deprecated ts file was remaining (`emailing.ts`). I did some cleaning
and tested the requests to Brevo and TEE backend to be sure I didn't
break anything...