refactor: remove legacy conditions (#133)
The aim of this PR is to remove legacy conditions handling completely,
so that `publicodes` is the unique mechanism of filtrering.

My observation is that there are some legacy conditions that are not
used anymore: `project_needs` and `project_status` are related to
questions that have been removed from the questionnaire.

`project_sectors` and `user_goals` however are still used, and imply a
little paradigm shift : filtering on sectors does not imply eligibility
rules, but rather content curation. As such, it is a little bit
unsettling to put the rules under the name `éligibilté` in publicodes.

Eventually, we will not want to mix up eligibility rules and content
curation in the same data file, but for now the most straightforward way
I found to mark the difference is to put a new mechanism called
`afficher dispositif si` which is a superset of `éligibilité` (renamed
`règles d'éligibilité`).

In order to factorize the expected interface between typescript and
publicodes, a `common/interface.yaml` is prepended to every `publicodes`
attribute of program data.

- As a result, the `publicodes` attribute in every "program" data file
is not self-standing anymore. However, it becomes self-standing when
prepended with "interface.yaml".
- In practice, data is consolidated in the `buildJSONOutput.ts`
pipeline, for frontend availability.
### Other changes

- The JSON Schema has been migrated from "./packages/backend/data" to
"./packages/data/schemas"
- Typescript type for `Program` is (partly) generated from json schema
to avoid duplicate type definitions, using
[`json-schema-to-typescript`](https://www.npmjs.com/package/json-schema-to-typescript)
- Conditionnaly dependent fields ("coût de l'accompagnement", "montant
de l'aide" etc.) in json-schema has been rewritten to be handled well by
`json-schema-to-typescript`, and for better readability.
- The "test-data" script now lives in the "data" package, and runs
"./packages/data/tests/test-program-data.sh"

---------

Co-authored-by: Julien Paris <jparis.py@gmail.com>