hotfix: filtering problem in "Je ne sais pas par où commencer"  (#151)
To avoid unnoticed filtering problems, publicodes would require to have
all variables defined when evaluating a rule, or otherwise throw an
error.

However, if filtering problems are not noticed in preproduction, then
the final user gets the error, which is not the best ID. In particular,
we missed that in the "Je ne sais pas par où commencer" route was not
defining the `questionnaire . objectif prioritaire` property, used to
filter some programs.

This hotfix allows to have undefined publicodes rules. The introduced
behavior is to keep the programs with undefined `afficher à
l'utilisateur si` rule.