feat: add TeeBanner component to ProgramList
A new component, TeeBanner, has been created and imported into the ProgramList component. TeeBanner includes title and description props for better modularity and is going to be used for displaying banners on program list pages. Additional styling classes have been applied to ensure it fits in with the overall design.