feat: update components to use dynamic project data
Replaced the static project data with dynamic project data from the project store in different components such as ResultProjectList and ResultList. This change also included creating a new store for project data and incorporating its use in API calls and presentation components. This ensures that the project list displayed to the user is always up-to-date.