feat: add function to filter projects by objective
The commit introduces a new function named 'getProjectsByObjective' in the project store that allows filtering of projects based on a specific objectiveType in the Vue project. The 'ResultProjectList' and 'ResultList' components are updated to use the newly filtered projects instead of using all the projects. A new utils file 'projectFilters' has been added to assist the filtering by theme.