feat: add eligibility filter to project listing
A new function has been added to the project filters to filter projects by eligibility. This new filter has been incorporated into the `getProjectsByObjectiveAndEligibility` function in the Project store. The project listing now considers the eligibility of a project along with its alignment with the user's objective to provide a more refined and accurate result list.